event/cnxk: fix getwork mode devargs parsing
CN10K can support up to three types of getwork modes, use parse value
function to parse the device arguments.

Fixes: 5fb651552c79 ("event/cnxk: fix SSO and TIM argument parsing")
Cc: stable@dpdk.org

+++ drivers/event/cnxk/cnxk_eventdev.h
@@ -119,7 +119,7 @@ struct cnxk_sso_evdev {
 	/* CN9K */
 	uint8_t dual_ws;
 	/* CN10K */
-	uint8_t gw_mode;
+	uint32_t gw_mode;
 	uint16_t stash_cnt;
 	struct cnxk_sso_stash *stash_parse_data;
 } __rte_cache_aligned;
